requirements.txt
streamlit
snowflake-connector-python
pandas
plotly
streamlit-folium
snowflake-snowpark-python